+++ fxmanifest.lua
@@ -8,7 +8,7 @@ author "Giga Scripts"
 
 client_scripts { "client.lua" }
 server_scripts { "@oxmysql/lib/MySQL.lua", "server.lua" }
-shared_scripts { '@ox_lib/init.lua', "config.lua", "framework.lua", "bridge/imports.lua" }
+shared_scripts { '@ox_lib/init.lua', "config.lua", "framework.lua", "shared/vehicle_images.lua", "bridge/imports.lua" }
 
 files {
     "ui/dist/**/*",